updated!!! 15-11-2025

+++ lib/views/widgets/text_style.dart
@@ -2,10 +2,9 @@ import 'package:emartseller/const/const.dart';
 import 'package:velocity_x/velocity_x.dart';
 
 Widget normalText({text,color= Colors.white,size=14.0}){
-  return "$text".text.color(color).size(size).make();
+  return "$text".text.color(color).size(size.toDouble()).make();
 }
 
 Widget boldText({text,color = Colors.white,size =14.0}){
-  return "$text".text.semiBold.color(color).size(size).make();
+  return "$text".text.semiBold.color(color).size(size.toDouble()).make();
 }
-
